==> Notebooks/01. DataCollection.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

#!pip install git+https://github.com/JustAnotherArchivist/snscrape.git
!pip install snscrape
!pip install tqdm

import snscrape.modules.twitter as sntwitter
import pandas as pd
from IPython.display import display
import datetime
import os
from tqdm import tqdm

x = datetime.datetime(2022, 2, 24) #24th Feb 2022
nowx = datetime.datetime(2022, 10, 31) #31th Oct 2022
while x != nowx:
    y=x+datetime.timedelta(days=1)
    since_month = x.strftime("%m")
    since_day = x.strftime("%d")
    until_month = y.strftime("%m")
    until_day = y.strftime("%d")
    filename = '/content/drive/MyDrive/Research/SentimentAnalysis/RawTweets/tweets-'+since_day+'-'+since_month+'-2022.csv'
    isFile = os.path.isfile(filename)
    if isFile:
        print(filename+" already exists...")
        x=y
        continue
        
    query = "Ukraine Russia (War OR Conflict) lang:en until:2022-"+until_month+"-"+until_day+" since:2022-"+since_month+"-"+since_day
    print("Collecting Tweets: "+since_day+"-"+since_month+"-2022")
    tweets = []
    limit = 10000
    pbar = tqdm(total=limit)
    for tweet in sntwitter.TwitterSearchScraper(query).get_items():
        if len(tweets) == limit:
            break
        else:
            tweets.append(tweet)
            pbar.update(1)
    pbar.close()  
    df = pd.DataFrame(tweets)
    display(df.shape)
    df.to_csv(filename)
    x=y
