==> Notebooks/03. DataLabeling(Hybrid Ensemble).ipynb <==
# **Data Labeling**
Datasets will be labeled using the Majority Voting or Averaging based Ensemble method depending on the following 4 pre-trained models and libraries.

1. Twitter-roBERTa-base (https://arxiv.org/abs/2010.12421)
2. XLM-T (https://arxiv.org/abs/2104.12250)
3. VADER (https://ojs.aaai.org/index.php/ICWSM/article/view/14550)
4. TextBlob

Primarily Majority Voting will be used. Is there is a tie in the top voting count, Averaging will be use to make the decision.

!pip install transformers
!pip install sentencepiece
!pip install vaderSentiment
!pip install textblob


from transformers import AutoModelForSequenceClassification
from transformers import TFAutoModelForSequenceClassification
from transformers import AutoTokenizer, AutoConfig
import numpy as np
from scipy.special import softmax

from transformers import pipeline

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

from textblob import TextBlob

def preprocess(text):
    text = str(text)
    new_text = []
    for t in text.split(" "):
        t = '@user' if t.startswith('@') and len(t) > 1 else t
        t = 'http' if t.startswith('http') else t
        new_text.append(t)
    return " ".join(new_text)

# **Twitter-roBERTa-base(2022)**

MODEL = f"cardiffnlp/twitter-roberta-base-sentiment-latest"
tokenizer = AutoTokenizer.from_pretrained(MODEL)
config = AutoConfig.from_pretrained(MODEL)
model = AutoModelForSequenceClassification.from_pretrained(MODEL)

def sentiment_trb(text):
    text = preprocess(text)
    encoded_input = tokenizer(text, return_tensors='pt')
    output = model(**encoded_input)
    scores = output[0][0].detach().numpy()
    scores = softmax(scores)
    ranking = np.argsort(scores)
    ranking = ranking[::-1]
    return config.id2label[ranking[0]]

# **XLM-T**

model_path = "cardiffnlp/twitter-xlm-roberta-base-sentiment"
sentiment_task = pipeline("sentiment-analysis", model=model_path, tokenizer=model_path)

def sentiment_xlmt(text):
  return sentiment_task(text)[0]["label"]

# **VADER**

def sentiment_vader(text):
    sid_obj = SentimentIntensityAnalyzer()
    sentiment_dict = sid_obj.polarity_scores(text)
    compound = sentiment_dict['compound']
    if sentiment_dict['compound'] >= 0.05:
        overall_sentiment = "positive"
    elif sentiment_dict['compound'] <= -0.05:
        overall_sentiment = "negative"
    else :
        overall_sentiment = "neutral"
    return overall_sentiment

# **TextBlob**

def sentiment_texblob(text):
    classifier = TextBlob(text)
    polarity = classifier.sentiment.polarity  
    if polarity >= 0.05:
        overall_sentiment = "positive"
    elif polarity <= -0.05:
        overall_sentiment = "negative"
    else :
        overall_sentiment = "neutral"
    return overall_sentiment

# **Ensemble Method**

!pip install tqdm

import pandas as pd
import datetime
import os
from tqdm import tqdm
pd.options.mode.chained_assignment = None  # default='warn'

x = datetime.datetime(2022, 2, 24) #24th Feb 2022
nowx = datetime.datetime(2022,3,20) #31th Oct 2022

while x != nowx:
  month = x.strftime("%m")
  day = x.strftime("%d")
  filename = '/content/drive/MyDrive/FSTweets/fs-tweets-'+day+'-'+month+'-2022.csv'
  isFile = os.path.isfile(filename)
  if isFile:
        desfilename = '/content/drive/MyDrive/LabeledTweets/labeled-tweets-'+day+'-'+month+'-2022.csv'
        desFile = os.path.isfile(desfilename)
        if desFile:
            print(desfilename+" already exists...")
            x=x+datetime.timedelta(days=1)
            continue
        df = pd.read_csv(filename)

        labels = []
        limit = df.shape[0]
        pbar = tqdm(total=limit, position=0, leave=True)
        for idx in df.index:
            text = df['rawContent'][idx]
            text = preprocess(text)

            voting_cnt = {
                "positive": 0,
                "negative": 0,
                "neutral": 0,
            }
            
            vote = sentiment_xlmt(text)
            voting_cnt[vote] = voting_cnt[vote] + 1
            vote = sentiment_vader(text)
            voting_cnt[vote] = voting_cnt[vote] + 1
            vote = sentiment_texblob(text)
            voting_cnt[vote] = voting_cnt[vote] + 1

            mx_vote = 0

            for ct in voting_cnt.values():
                mx_vote = max(mx_vote,ct)

            mx_vote_ct=0

            for label in voting_cnt:
                if voting_cnt[label] == mx_vote:
                    mx_vote_ct = mx_vote_ct + 1

            if mx_vote_ct == 1:
                for label in voting_cnt:
                    if voting_cnt[label] == mx_vote:
                        final_label = label
                        break
            else:
                vote = sentiment_trb(text)
                voting_cnt[vote] = voting_cnt[vote] + 1

                mx_vote = 0

                for ct in voting_cnt.values():
                    mx_vote = max(mx_vote,ct)

                mx_vote_ct=0

                for label in voting_cnt:
                    if voting_cnt[label] == mx_vote:
                        mx_vote_ct = mx_vote_ct + 1

                if mx_vote_ct == 1:
                    for label in voting_cnt:
                        if voting_cnt[label] == mx_vote:
                            final_label = label
                            break
                else:
                    avg_prediction = 0
                    for label in voting_cnt:
                        if label == 'positive':
                            weight = 1
                        elif label == 'neutral':
                            weight = 0
                        else:
                            weight = -1
                        avg_prediction = avg_prediction + (weight * voting_cnt[label])

                    avg_prediction = avg_prediction / 4

                    if(avg_prediction > 0):
                        final_label = 'positive'
                    elif(avg_prediction < 0):
                        final_label = 'negative'
                    else:
                        final_label = 'neutral'

            labels.append(final_label)
            pbar.update(1)
            
        pbar.close()
        df['label'] = labels
        display(df)
        df.to_csv('/content/drive/MyDrive/LabeledTweets/labeled-tweets-'+day+'-'+month+'-2022.csv',index=False)
        print(f"Done: "+day+"/"+month)
        x=x+datetime.timedelta(days=1)
