==> Notebooks/09. DatasetFiltering(Final Dataset).ipynb <==
import pandas as pd

df = pd.read_csv("/content/drive/MyDrive/Research/SentimentAnalysis/LabeledTweetsBinary/label-tweets-all-with-alt.csv")
df

df = df.drop(columns={"hashtags","id"})
df

df_neu = df[df['label']=='neutral']
df_neu_neg = df_neu[df_neu['label_alt']=='negative']
df_neu_pos = df_neu[df_neu['label_alt']=='positive']

df_opp = df[df['label']!='neutral']
df_opp = df_opp[df_opp['label']!=df_opp['label_alt']]
df_pos_neg = df_opp[df_opp['label_alt']=='negative']
df_neg_pos = df_opp[df_opp['label_alt']=='positive']

texts = df_pos_neg['rawContent'].to_list()

for i in range(100):
    print(f"{i+1}: {texts[i]}")

texts = df_neg_pos['rawContent'].to_list()

for i in range(100):
    print(f"{i+1}: {texts[i]}")

texts = df_neu_neg['rawContent'].to_list()

for i in range(100):
    print(f"{i+1}: {texts[i]}")

texts = df_neu_pos['rawContent'].to_list()

for i in range(100):
    print(f"{i+1}: {texts[i]}")

df = df[df['label']==df['label_alt']]
df

df=df.drop(columns={"label_alt"})
df

df=df.rename(columns={"rawContent":"text"})
df

df.to_csv('/content/drive/MyDrive/Research/SentimentAnalysis/LabeledTweetsBinary/label-tweets-final.csv',index=False)
