==> Notebooks/10. FinalDatasetSmall.ipynb <==
import pandas as pd
df = pd.read_csv("/content/drive/MyDrive/Research/SentimentAnalysis/LabeledTweetsBinary/label-tweets-final.csv")
df

df_pos = df[df['label']=='positive']
df_neg = df[df['label']=='negative']

pos_ct = len(df_pos['label'].to_list())
neg_ct = len(df_neg['label'].to_list())

mx_len = min(pos_ct,neg_ct)
print(mx_len)

g = df.groupby('label')
flags = (g.cumcount() + 1) <= mx_len
flags
df = df.loc[flags].sort_index()
df

df.to_csv('/content/drive/MyDrive/Research/SentimentAnalysis/LabeledTweetsBinary/label-tweets-final-balanced.csv',index=False)
